Move the projection-dsl example to example-projection
Relocates example/domain/projection-dsl to example/projection/projection-dsl
and renames its package from org.occurrent.example.domain.projection to
org.occurrent.example.projection.dsl, matching the example-projection group's
naming (for example example-projection-global-position-catchup uses
org.occurrent.example.projection.globalposition). Retargets the module pom to
the example-projection parent and rewires both aggregator poms. The CI shards
are path-based, so the module moves to the example-projection shard with no
workflow change.
